added some comments

+++ src/core/registry.rs
@@ -1,7 +1,7 @@
 use super::meta::MetaMarketData;
 
 pub trait Registrable {
-    fn registry_id(&self) -> Option<u64>;
-    fn register_id(&mut self, id: u64);
+    fn registry_id(&self) -> Option<usize>;
+    fn register_id(&mut self, id: usize);
     fn meta_market_data(&self) -> MetaMarketData;
 }
